+++ src/main/java/com/projectileoverride/RegionIds.java
@@ -1,0 +1,10 @@
+package com.projectileoverride;
+
+public final class RegionIds
+{
+	public static final int[] NONE = null;
+
+	public static final int[] DAGGANOTH_KINGS = new int[] {11588, 11589}; // Regular and slayer instance.
+
+	private RegionIds() {}
+}