+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this plugin will be documented in this file.
 
 ----------------------------------------------------------------------
+## [1.0.2] - 2025-08-31
+### Added
+- Added Doom rocks, Zebak and Zebak rocks.
+### Fixed
+- Fixed an issue where the Dagganoth Kings override would override all water wave projectiles.
+
 ## [1.0.1] - 2025-08-12
 ### Added
 - Added Cerberus, Dagganoth Kings and Zulrah.